Fix warning.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@19500 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/generic/progdlgg.cpp
@@ -337,7 +337,7 @@ wxProgressDialog::Update(int value, const wxString& newmsg)
     if ( (m_elapsed || m_remaining || m_estimated) && (value != 0) )
     {
         unsigned long elapsed = wxGetCurrentTime() - m_timeStart;
-        unsigned long estimated = ( (double) elapsed * m_maximum ) / ((double)value) ;
+        unsigned long estimated = (unsigned long)(( (double) elapsed * m_maximum ) / ((double)value)) ;
         unsigned long remaining = estimated - elapsed;
 
         SetTimeLabel(elapsed, m_elapsed);
